--- src/lib/server/_shared/constants.ts
@@ -1,1 +1,0 @@
-export const NUM_DAYS_IN_SCHED = 21;
